backend/requirements.txt
# Backend Dependencies for OCR Document Processing

# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# OCR and Image Processing
pytesseract==0.3.13
opencv-python==4.10.0.84
Pillow==11.0.0
pdf2image==1.17.0

# NLP and Entity Recognition
spacy==3.8.7
python-dateutil==2.9.0
google-generativeai==0.8.3

# Machine Learning (for fraud detection)
scikit-learn==1.6.1
pandas==2.2.3

# Data Processing
pydantic==2.10.3
pydantic-settings==2.6.1

# Utilities
python-dotenv==1.0.1
numpy==2.2.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and Monitoring
loguru==0.7.3

# Testing
pytest==8.3.4
httpx==0.28.1
